--- packages/cms/myrun.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-NODE_ENV=test yarn dev
